# labs/lab03-rest-api/starter/openapi.yaml
openapi: 3.0.3
info:
  title: Lab 3 Item API
  version: 1.0.0
  description: A simple REST API for managing an in-memory collection of items.

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required:
                  - status

  /items:
    get:
      summary: Get all items
      responses:
        "200":
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /items/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get an item by ID
      responses:
        "200":
          description: The requested item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an item by ID
      responses:
        "204":
          description: Item deleted successfully
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: keyboard
        quantity:
          type: integer
          example: 10
      required:
        - id
        - name
        - quantity
    ItemInput:
      type: object
      properties:
        name:
          type: string
          example: monitor
        quantity:
          type: integer
          example: 4
      required:
        - name
        - quantity
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Item not found
      required:
        - error